FCFS Algorithm, When Arrival Time of All the Processes is Same
In this code I am provide the facility to enter the CPU burst time of all the processes and consider the arrival time of all the processes is same. So this particular code help you to understand how the FCFS Algorithm work when the arrival time of all the processes is same i.e. equal to zero.